refactor based on modular architecture
