+++ src/demos/config.js
@@ -18,11 +18,39 @@ import {
   HOMER_SELECT_CARD,
   HOMER_LOOKUP_REFERENCE,
   SET_IMAGE_URL,
-  TOGGLE_LEFT_SIDEBAR_VISIBILITY,
-  TOGGLE_RIGHT_SIDEBAR_VISIBILITY,
+  CHANGE_SIDEBAR_VISIBILITY,
 } from './constants';
 
 import cards from './homer';
+
+const getDefaultState = () => {
+  return {
+    rightOpen: true,
+    leftOpen: true,
+    leftVisible: true,
+    rightVisible: true,
+    widgets: {
+      mainWidget: null,
+      right: [],
+      left: [],
+    },
+    passageText: '',
+    selectedWords: [],
+    readerText: {},
+    selectedLemmas: [],
+    readerTextSize: 'md',
+    readerTextWidth: 'normal',
+
+    interlinear: false,
+    books: [],
+    book: null,
+    passage: null,
+    word: null,
+    selectedCard: null,
+    cards,
+    imageURL: '',
+  };
+};
 
 const parseHomerReference = (ref) => {
   const parts = ref.split('.');
@@ -34,32 +62,7 @@ const parseHomerReference = (ref) => {
 
 export default function createStore() {
   return {
-    state: {
-      rightOpen: true,
-      leftOpen: true,
-      leftVisible: true,
-      rightVisible: true,
-      widgets: {
-        mainWidget: null,
-        right: [],
-        left: [],
-      },
-      passageText: '',
-      selectedWords: [],
-      readerText: {},
-      selectedLemmas: [],
-      readerTextSize: 'md',
-      readerTextWidth: 'normal',
-
-      interlinear: false,
-      books: [],
-      book: null,
-      passage: null,
-      word: null,
-      selectedCard: null,
-      cards,
-      imageURL: '',
-    },
+    state: getDefaultState,
     getters: {
       // eslint-disable-next-line arrow-parens
       getChunks: (state) => (start, end) => {
@@ -175,11 +178,13 @@ export default function createStore() {
       [SET_IMAGE_URL]: (state, url) => {
         state.imageURL = url;
       },
-      [TOGGLE_LEFT_SIDEBAR_VISIBILITY]: (state) => {
-        state.leftVisible = !state.leftVisible;
-      },
-      [TOGGLE_RIGHT_SIDEBAR_VISIBILITY]: (state) => {
-        state.rightVisible = !state.rightVisible;
+      [CHANGE_SIDEBAR_VISIBILITY]: (state, { side, bool }) => {
+        if (side === 'left') {
+          state.leftVisible = bool;
+        }
+        if (side === 'right') {
+          state.rightVisible = bool;
+        }
       },
     },
     actions: {
@@ -260,11 +265,8 @@ export default function createStore() {
       [SET_IMAGE_URL]: ({ commit }, { url }) => {
         commit(SET_IMAGE_URL, url);
       },
-      [TOGGLE_LEFT_SIDEBAR_VISIBILITY]: ({ commit }) => {
-        commit(TOGGLE_LEFT_SIDEBAR_VISIBILITY);
-      },
-      [TOGGLE_RIGHT_SIDEBAR_VISIBILITY]: ({ commit }) => {
-        commit(TOGGLE_RIGHT_SIDEBAR_VISIBILITY);
+      [CHANGE_SIDEBAR_VISIBILITY]: ({ commit }, { side, bool }) => {
+        commit(CHANGE_SIDEBAR_VISIBILITY, { side, bool });
       },
     },
   };
