+++ src/demos/constants.js
@@ -15,5 +15,4 @@ export const MORPHGNT_SET_SELECTED_WORD = 'setSelectedWord';
 export const HOMER_SELECT_CARD = 'homerSelectCard';
 export const HOMER_LOOKUP_REFERENCE = 'homerLookupReference';
 export const SET_IMAGE_URL = 'setImageURL';
-export const TOGGLE_LEFT_SIDEBAR_VISIBILITY = 'toggleLeftSidebarVisibility';
-export const TOGGLE_RIGHT_SIDEBAR_VISIBILITY = 'toggleRightSidebarVisibility';
+export const CHANGE_SIDEBAR_VISIBILITY = 'changeSidebarVisibility';
