+++ src/skeleton/constants.js
@@ -5,5 +5,4 @@ export const REMOVE_LEFT_WIDGET = 'removeLeftWidget';
 export const ADD_RIGHT_WIDGET = 'addRightWidget';
 export const REMOVE_RIGHT_WIDGET = 'removeRightWidget';
 export const CHANGE_MAIN_WIDGET = 'changeMainWidget';
-export const TOGGLE_LEFT_SIDEBAR_VISIBILITY = 'toggleLeftSidebarVisibility';
-export const TOGGLE_RIGHT_SIDEBAR_VISIBILITY = 'toggleRightSidebarVisibility';
+export const CHANGE_SIDEBAR_VISIBILITY = 'changeSidebarVisibility';
